Added information to get headers from a csv file, make/show a 2d array of current database, collect informationn about the database like Table names and column names in each table. Added the ability to search for a header in the database and find where it is located. Made a function to start the searching for the headers in the database.
Need to add the following:
Cross referencing of CSV headers to Database table Columns.
Making of mysql insert statements to Database using the information collected
from the findColumn information.
Finding a way to double check if information is already stored in database.
Finding a way to update information if needed when patient is found in database.
Last but not least insert the statements into the database for storage.

Joel Buehler 3/12/11